+++ packages/nhsuk-frontend/src/nhsuk/components/legend/macro-options.mjs
@@ -32,7 +32,14 @@ const options = {
   isPageHeading: {
     type: 'boolean',
     required: false,
-    description: 'Whether the legend also acts as the heading for the page.'
+    description:
+      'Whether the legend also acts as the heading for the page. Defaults to `true` when `headingLevel` is provided.'
+  },
+  headingLevel: {
+    type: 'integer',
+    required: false,
+    description:
+      'Optional legend heading level. Defaults to `1` when `isPageHeading` is `true`.'
   },
   size: {
     type: 'string',